+++ src/MR.EntityFrameworkCore.KeysetPagination.Analyzers/DiagnosticDescriptors.cs
@@ -11,14 +11,14 @@ public static class DiagnosticDescriptors
 
 	// https://github.com/dotnet/roslyn-analyzers/issues/5828
 #pragma warning disable IDE0090 // Use 'new(...)'
-	public static readonly DiagnosticDescriptor KeysetPagination1000_KeysetContainsNullableProperty = new DiagnosticDescriptor(
+	public static readonly DiagnosticDescriptor KeysetPagination1000_KeysetContainsNullableColumn = new DiagnosticDescriptor(
 		"KeysetPagination1000",
-		"Keyset contains a nullable property",
-		"Unsupported nullable property '{0}' in the keyset",
+		"Keyset column may be null",
+		"Keyset column may be null",
 		Categories.Usage,
 		DiagnosticSeverity.Error,
 		isEnabledByDefault: true,
-		description: "Nullable properties are not supported in the keyset.",
+		description: "Nullable columns are not supported in the keyset.",
 		helpLinkUri: "https://github.com/mrahhal/MR.EntityFrameworkCore.KeysetPagination/blob/main/docs/diagnostics.md#KeysetPagination1000");
 #pragma warning restore IDE0090 // Use 'new(...)'
 }