+++ src/MR.EntityFrameworkCore.KeysetPagination/Exceptions.cs
@@ -3,7 +3,7 @@
 public class KeysetPaginationIncompatibleObjectException : Exception
 {
 	public KeysetPaginationIncompatibleObjectException(string message)
-		: base(message + " Refer to the following document for more info: https://github.com/mrahhal/MR.EntityFrameworkCore.KeysetPagination/blob/main/docs/loose-typing.md")
+		: base(message + " Refer to the following document for more info on loose typing: https://github.com/mrahhal/MR.EntityFrameworkCore.KeysetPagination/blob/main/docs/loose-typing.md")
 	{
 	}
 }