+++ test/MR.EntityFrameworkCore.KeysetPagination.Tests/DatabaseCollection.cs
@@ -1,0 +1,15 @@
+﻿using Xunit;
+
+namespace MR.EntityFrameworkCore.KeysetPagination;
+
+[CollectionDefinition(Name)]
+public class SqlServerDatabaseCollection : ICollectionFixture<SqlServerDatabaseFixture>
+{
+	public const string Name = nameof(SqlServerDatabaseCollection);
+}
+
+[CollectionDefinition(Name)]
+public class SqliteDatabaseCollection : ICollectionFixture<SqliteDatabaseFixture>
+{
+	public const string Name = nameof(SqliteDatabaseCollection);
+}