+++ test/MR.EntityFrameworkCore.KeysetPagination.Tests/TestModels/GuidModel.cs
@@ -1,8 +1,0 @@
-﻿namespace MR.EntityFrameworkCore.KeysetPagination.TestModels;
-
-public class GuidModel
-{
-	public Guid Id { get; set; }
-
-	public DateTime Created { get; set; }
-}