+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+
+### Improved
+
+- Expression adapting ([#37](https://github.com/mrahhal/MR.EntityFrameworkCore.KeysetPagination/issues/37) by [@mrahhal](https://github.com/mrahhal))
+
+[**Full diff**](https://github.com/mrahhal/MR.EntityFrameworkCore.KeysetPagination/compare/v1.2.0...HEAD)
+
 ## 1.2.0 - 2022-11-03
 
 ### Improved